UInt -> Int

+++ Realm/RealmSwift/List.swift
@@ -23,7 +23,7 @@ public class ListBase: RLMListBase, Printable {
     // and it has to be defined as @objc override, which can't be done in a
     // generic class.
     @objc public override var description: String { return _rlmArray.description }
-    public var count: UInt { return _rlmArray.count }
+    public var count: Int { return Int(_rlmArray.count) }
 }
 
 public final class List<T: Object>: ListBase, SequenceType {
@@ -48,26 +48,26 @@ public final class List<T: Object>: ListBase, SequenceType {
 
     // MARK: Index Retrieval
 
-    public func indexOf(object: T) -> UInt? {
+    public func indexOf(object: T) -> Int? {
         return notFoundToNil(_rlmArray.indexOfObject(unsafeBitCast(object, RLMObject.self)))
     }
 
-    public func indexOf(predicate: NSPredicate) -> UInt? {
+    public func indexOf(predicate: NSPredicate) -> Int? {
         return notFoundToNil(_rlmArray.indexOfObjectWithPredicate(predicate))
     }
 
-    public func indexOf(predicateFormat: String, _ args: CVarArgType...) -> UInt? {
+    public func indexOf(predicateFormat: String, _ args: CVarArgType...) -> Int? {
         return notFoundToNil(_rlmArray.indexOfObjectWhere(predicateFormat, args: getVaList(args)))
     }
 
     // MARK: Object Retrieval
 
-    public subscript(index: UInt) -> T {
+    public subscript(index: Int) -> T {
         get {
-            return _rlmArray[index] as T
+            return _rlmArray[UInt(index)] as T
         }
         set {
-            return _rlmArray[index] = newValue
+            return _rlmArray[UInt(index)] = newValue
         }
     }
 
@@ -120,12 +120,12 @@ public final class List<T: Object>: ListBase, SequenceType {
         }
     }
 
-    public func insert(object: T, atIndex index: UInt) {
-        _rlmArray.insertObject(unsafeBitCast(object, RLMObject.self), atIndex: index)
+    public func insert(object: T, atIndex index: Int) {
+        _rlmArray.insertObject(unsafeBitCast(object, RLMObject.self), atIndex: UInt(index))
     }
 
-    public func remove(index: UInt) {
-        _rlmArray.removeObjectAtIndex(index)
+    public func remove(index: Int) {
+        _rlmArray.removeObjectAtIndex(UInt(index))
     }
 
     public func remove(object: T) {
@@ -142,16 +142,16 @@ public final class List<T: Object>: ListBase, SequenceType {
         _rlmArray.removeAllObjects()
     }
     
-    public func replace(index: UInt, object: T) {
-        _rlmArray.replaceObjectAtIndex(index, withObject: unsafeBitCast(object, RLMObject.self))
+    public func replace(index: Int, object: T) {
+        _rlmArray.replaceObjectAtIndex(UInt(index), withObject: unsafeBitCast(object, RLMObject.self))
     }
 
     // MARK: Private stuff
 
-    private func notFoundToNil(index: UInt) -> UInt? {
+    private func notFoundToNil(index: UInt) -> Int? {
         if index == UInt(NSNotFound) {
             return nil
         }
-        return index
+        return Int(index)
     }
 }
